toolkit/themes/pinstripe/tests/test_bug371080.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="http://people.mozilla.com/~sayrer/tests/SimpleTest/test.css" type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=371080
-->
<window title="Mozilla Bug 371080"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  align="start">
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>      

<body  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=371080">Mozilla Bug 371080</a>
</body>

<!-- This Mac OS X native button should be 68 pixels wide from
     the dark border line on the left to the dark border line on
     the right. It should be 21 pixels tall from the dark border
     line on the top to the dark border line on the bottom -->
<button class="bigbutton"
        label="OK"
        width="5"
        id="button1"
        height="5"/>
<!-- This Mac OS X native button should be 68 pixels wide from
     the dark border line on the left to the dark border line on
     the right. It should be 18 pixels tall from the dark border
     line on the top to the dark border line on the bottom -->
<button class="smallbutton"
        style="-moz-appearance: button-small;"
        label="OK"
        width="5"
        id="button2"
        height="5"/>

<script class="testbody" type="application/javascript">
<![CDATA[

/** Test for Bug 371080 **/
SimpleTest.waitForExplicitFinish();

function checkButtonDimensions() {
  is($("button1").boxObject.width, 68, "mac button has correct width");
  is($("button1").boxObject.height, 22, "mac button has correct height");
  is($("button2").boxObject.width, 68, "mac small button has correct width");
  is($("button2").boxObject.height, 19, "mac small button has correct height");
}

addLoadEvent(checkButtonDimensions);
addLoadEvent(SimpleTest.finish);

]]>
</script>
</window>
